+++ packages/interop/csv/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @spare/csv
 
+## 0.16.16
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/table-margin@0.16.16
+  - @spare/table-padder@0.16.16
+
+## 0.16.15
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/table-margin@0.16.15
+  - @spare/table-padder@0.16.15
+
+## 0.16.14
+
+### Patch Changes
+
+- update dependencies
+- Updated dependencies
+  - @spare/table-margin@0.16.14
+  - @spare/table-padder@0.16.14
+
+## 0.16.13
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/table-margin@0.16.13
+  - @spare/table-padder@0.16.13
+
 ## 0.16.12
 
 ### Patch Changes