+++ packages/interop/deco-samples/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @spare/deco-samples
 
+## 0.16.16
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-flat@0.16.16
+  - @spare/enum-brackets@0.16.16
+  - @spare/preset-deco@0.16.16
+  - @spare/node@0.16.16
+  - @spare/deco-config@0.16.16
+
+## 0.16.15
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-flat@0.16.15
+  - @spare/enum-brackets@0.16.15
+  - @spare/preset-deco@0.16.15
+  - @spare/node@0.16.15
+  - @spare/deco-config@0.16.15
+
+## 0.16.14
+
+### Patch Changes
+
+- update dependencies
+- Updated dependencies
+  - @spare/deco-flat@0.16.14
+  - @spare/enum-brackets@0.16.14
+  - @spare/preset-deco@0.16.14
+  - @spare/node@0.16.14
+  - @spare/deco-config@0.16.14
+
+## 0.16.13
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/enum-brackets@0.16.13
+  - @spare/preset-deco@0.16.13
+  - @spare/deco-flat@0.16.13
+  - @spare/node@0.16.13
+  - @spare/deco-config@0.16.13
+
 ## 0.16.12
 
 ### Patch Changes