+++ packages/interop/verse/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @spare/verse
 
+## 0.16.16
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-entries@0.16.16
+  - @spare/deco-matrix@0.16.16
+  - @spare/deco-object@0.16.16
+  - @spare/deco-vector@0.16.16
+  - @spare/deco-samples@0.16.16
+  - @spare/enum-brackets@0.16.16
+  - @spare/preset-verse@0.16.16
+  - @spare/quote@0.16.16
+
+## 0.16.15
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-entries@0.16.15
+  - @spare/deco-matrix@0.16.15
+  - @spare/deco-object@0.16.15
+  - @spare/deco-vector@0.16.15
+  - @spare/deco-samples@0.16.15
+  - @spare/enum-brackets@0.16.15
+  - @spare/preset-verse@0.16.15
+  - @spare/quote@0.16.15
+
+## 0.16.14
+
+### Patch Changes
+
+- update dependencies
+- Updated dependencies
+  - @spare/deco-entries@0.16.14
+  - @spare/deco-matrix@0.16.14
+  - @spare/deco-object@0.16.14
+  - @spare/deco-vector@0.16.14
+  - @spare/deco-samples@0.16.14
+  - @spare/enum-brackets@0.16.14
+  - @spare/preset-verse@0.16.14
+  - @spare/quote@0.16.14
+
+## 0.16.13
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/enum-brackets@0.16.13
+  - @spare/preset-verse@0.16.13
+  - @spare/deco-samples@0.16.13
+  - @spare/deco-entries@0.16.13
+  - @spare/deco-matrix@0.16.13
+  - @spare/deco-object@0.16.13
+  - @spare/deco-vector@0.16.13
+  - @spare/quote@0.16.13
+
 ## 0.16.12
 
 ### Patch Changes