+++ packages/primitive/deco-date/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @spare/deco-date
 
+## 0.16.16
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/preset-deco@0.16.16
+  - @spare/deco-config@0.16.16
+
+## 0.16.15
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/preset-deco@0.16.15
+  - @spare/deco-config@0.16.15
+
+## 0.16.14
+
+### Patch Changes
+
+- update dependencies
+- Updated dependencies
+  - @spare/preset-deco@0.16.14
+  - @spare/deco-config@0.16.14
+
+## 0.16.13
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/preset-deco@0.16.13
+  - @spare/deco-config@0.16.13
+
 ## 0.16.12
 
 ### Patch Changes