+++ packages/recursive/deco-pale/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @spare/deco-pale
 
+## 0.16.16
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-object@0.16.16
+  - @spare/deco-vector@0.16.16
+  - @spare/deco-func@0.16.16
+  - @spare/enum-brackets@0.16.16
+  - @spare/preset-deco@0.16.16
+  - @spare/deco-config@0.16.16
+  - @spare/deco-util@0.16.16
+  - @spare/quote@0.16.16
+
+## 0.16.15
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-object@0.16.15
+  - @spare/deco-vector@0.16.15
+  - @spare/deco-func@0.16.15
+  - @spare/enum-brackets@0.16.15
+  - @spare/preset-deco@0.16.15
+  - @spare/deco-config@0.16.15
+  - @spare/deco-util@0.16.15
+  - @spare/quote@0.16.15
+
+## 0.16.14
+
+### Patch Changes
+
+- update dependencies
+- Updated dependencies
+  - @spare/deco-object@0.16.14
+  - @spare/deco-vector@0.16.14
+  - @spare/deco-func@0.16.14
+  - @spare/enum-brackets@0.16.14
+  - @spare/preset-deco@0.16.14
+  - @spare/deco-config@0.16.14
+  - @spare/deco-util@0.16.14
+  - @spare/quote@0.16.14
+
+## 0.16.13
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/enum-brackets@0.16.13
+  - @spare/preset-deco@0.16.13
+  - @spare/deco-object@0.16.13
+  - @spare/deco-vector@0.16.13
+  - @spare/deco-func@0.16.13
+  - @spare/deco-config@0.16.13
+  - @spare/deco-util@0.16.13
+  - @spare/quote@0.16.13
+
 ## 0.16.12
 
 ### Patch Changes