+++ packages/recursive/deco/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @spare/deco
 
+## 0.16.16
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-date@0.16.16
+  - @spare/deco-func@0.16.16
+  - @spare/node@0.16.16
+  - @spare/deco-colors@0.16.16
+
+## 0.16.15
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-date@0.16.15
+  - @spare/deco-func@0.16.15
+  - @spare/node@0.16.15
+  - @spare/deco-colors@0.16.15
+
+## 0.16.14
+
+### Patch Changes
+
+- update dependencies
+- Updated dependencies
+  - @spare/deco-date@0.16.14
+  - @spare/deco-func@0.16.14
+  - @spare/node@0.16.14
+  - @spare/deco-colors@0.16.14
+
+## 0.16.13
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-date@0.16.13
+  - @spare/deco-func@0.16.13
+  - @spare/node@0.16.13
+  - @spare/deco-colors@0.16.13
+
 ## 0.16.12
 
 ### Patch Changes