+++ packages/recursive/node/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @spare/node
 
+## 0.16.16
+
+### Patch Changes
+
+- minor fix
+
+## 0.16.15
+
+### Patch Changes
+
+- minor fix
+
+## 0.16.14
+
+### Patch Changes
+
+- update dependencies
+
+## 0.16.13
+
+### Patch Changes
+
+- minor fix
+
 ## 0.16.12
 
 ### Patch Changes